NINJA-81 change the field names for source/rendered
Just as I was typing up the previous commit message I realized I don't
actually want the unsafe value to be the default, so I changed them around and
now 'description' is the safe rendered HTML value and 'source' is the raw
source for editing. I realize this is probably less convenient. Sorry.